[Feat] 전역 공간에 놓여진 함수의 캡슐화
- 일부 UI 컴포넌트를 생성하는 함수들이 전역 공간에 놓여져 있어 이를 각
  UI클래스의 extension으로 묶어서 해결.
